110  total-number-spanning-trees-graph
If a graph is a complete graph with n vertices, then total number of spanning trees is n(n-2) where n is the number of nodes in the graph. In complete graph, the task is equal to counting different labeled trees with n nodes for which have Cayley’s formula.

What if graph is not complete? 

Follow the given procedure:

STEP 1: Create Adjacency Matrix for the given graph. 
STEP 2: Replace all the diagonal elements with the degree of nodes. For eg. element at (1,1) position of adjacency matrix will be replaced by the degree of node 1, element at (2,2) position of adjacency matrix will be replaced by the degree of node 2, and so on. 
STEP 3: Replace all non-diagonal 1’s with -1. 
STEP 4: Calculate co-factor for any element. 
STEP 5: The cofactor that you get is the total number of spanning tree for that graph. 
Consider the following graph:

 kirchoff-formula 

Adjacency Matrix for the above graph will be as follows:

 kirchoff-matrix

 After applying STEP 2 and STEP 3, adjacency matrix will look like

 kirchoff-theorem 

The co-factor for (1, 1) is 8. Hence total no. of spanning tree that can be formed is 8. 

NOTE: Co-factor for all the elements will be same. Hence we can compute co-factor for any element of the matrix. This method is also known as Kirchhoff’s Theorem. It can be applied to complete graphs also.

let’s see another example to solve these problems by making use of the Laplacian matrix.

A Laplacian matrix L,
where L[i, i] is the degree of node i and L[i, j] = -1 if there is an edge between nodes i and j, and otherwise L[i, j] = 0.

Kirchhoff’s theorem provides a way to calculate the number of spanning trees for a given graph as a determinant of a special matrix.

consider the following graph,


the undirected graph

All possible spanning trees are as follows


three spanning trees

In order to calculate the number of spanning trees, construct a Laplacian matrix L, where L[i, i] is the degree of node i and L[i, j] = ?1 if there is an edge between nodes i and j, and otherwise L[i, j] = 0. 

for the above graph, The Laplacian matrix will look like this


 

The number of spanning trees equals the determinant of a matrix.

The Determinant of a matrix that can be obtained when we remove any row and any column from L.

For example, if we remove the first row and column, the result will be,


 

The determinant is always the same, regardless of which row and column we remove from L.