185  print-all-the-duplicates-in-the-input-string
Given a string S, the task is to print all the duplicate characters with their occurrences in the given string.

Example:

Input: S = “geeksforgeeks”
Output:
e, count = 4
g, count = 2
k, count = 2
s, count = 2

Explanation: g occurs 2 times in the string, k occurs 2 times in the string, s occurs 2 times in the string while e occurs 4 times in the string rest of the characters have unique occurrences.

Find the duplicate characters in a string using Sorting
If we sort the string then all duplicates will come together in the string. Then, traverse the string from starting index to the ending index and check if the neighbor character is same then increment the count by 1.

Recommended Problem
Same characters in two strings
Show Topics 
Solve Problem
Easy
62.38%
9.1K
Steps to implement the above approach:

Sort the given string.
Loop through the sorted string to find the duplicates.
If the next character is the same as the current character then we keep on counting the occurrence of that char.
If the count is greater than one then we print the character and its count.